merge master back to v12 branch

+++ src/spork.h
@@ -56,7 +56,7 @@ extern std::map<int, CSporkMessage> mapSporksActive;
 extern CSporkManager sporkManager;
 
 void ProcessSpork(CNode* pfrom, std::string& strCommand, CDataStream& vRecv);
-int GetSporkValue(int nSporkID);
+int64_t GetSporkValue(int nSporkID);
 bool IsSporkActive(int nSporkID);
 void ExecuteSpork(int nSporkID, int nValue);
 void ReprocessBlocks(int nBlocks);
